+++ kttk/context.py
@@ -4,7 +4,7 @@ import pickle
 import ktrack_api
 from kttk import template_manager, user_manager, utils
 
-
+# todo use ArgumentError instead
 class InvalidEntityException(Exception):
 
     def __init__(self, type_missing, id_missing):
@@ -15,11 +15,10 @@ class InvalidEntityException(Exception):
 class InvalidStepException(Exception):
 
     def __init__(self, step):
-        super(InvalidStepException, self).__init__("Invalid step, {} is not a non-empty string, its {}!".format(step, type(step)))
+        super(InvalidStepException, self).__init__(
+            "Invalid step, {} is not a non-empty string, its {}!".format(step, type(step)))
 
 
-# todo make them contain only type and id
-# todo step should be taken from task if task is provided
 class Context(object):
     # todo add docs
     """
@@ -91,7 +90,7 @@ class Context(object):
         """
         if step is not None:
             if isinstance(step, str) or isinstance(step, unicode):
-                if len(step) >0:
+                if len(step) > 0:
                     return True
             raise InvalidStepException(step)
         else:
@@ -249,8 +248,6 @@ class Context(object):
 
         # make sure to query all fields from ktrack, because we might only have id and type
 
-
-
         if self.entity:
             entity = kt.find_one(self.entity['type'], self.entity['id'])
             avaible_tokens['code'] = entity['code']
@@ -320,3 +317,68 @@ class Context(object):
             _user = user
 
         return Context(_project, _entity, _step, _task, _workfile, _user)
+
+    def populate_context(self):
+        """
+        Returns a PopulatedContext context instance based on this context
+        :return:
+        """
+        return PopulatedContext(project=self.project,
+                                entity=self.entity,
+                                step=self.step,
+                                task=self.task,
+                                workfile=self.workfile,
+                                user=self.user)
+
+
+class PopulatedContext(Context):
+
+    def __init__(self, project=None, entity=None, step=None, task=None, workfile=None, user=None):
+        """
+        Guarantes that provided entites are fully populated from database
+        :param project:
+        :param entity:
+        :param step:
+        :param task:
+        :param workfile:
+        :param user:
+        """
+        kt = ktrack_api.get_ktrack()
+        # project
+        self._validate_entity_dict(project)
+        if project:
+            self._project = kt.find_one('project', project['id'])
+        else:
+            self._project = None
+
+        # entity
+        self._validate_entity_dict(entity)
+        if entity:
+            self._entity = kt.find_one(entity['type'], entity['id'])
+        else:
+            self._entity = None
+
+        # step
+        self._validate_step(step)
+        self._step = step
+
+        # task
+        self._validate_entity_dict(task)
+        if task:
+            self._task = kt.find_one('task', task['id'])
+        else:
+            self._task = None
+
+        # workfile
+        self._validate_entity_dict(workfile)
+        if workfile:
+            self._workfile = kt.find_one('workfile', workfile['id'])
+        else:
+            self._workfile = None
+
+        # user
+        self._validate_entity_dict(user)
+        if user:
+            self._user = kt.find_one('user', user['id'])
+        else:
+            self._user = None